requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
black==23.12.1
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
dnspython==2.4.2
drf-spectacular==0.27.1
eventlet==0.34.3
flake8==7.0.0
flower==2.0.1
greenlet==3.0.3
humanize==4.9.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
